requirements.txt
requests>=2.31.0
requests-kerberos>=0.14.0
PyYAML>=6.0
click>=8.1.0
openai>=1.0.0

